refactor UseAgoraScreenShareProps / UseAgoraScreenShareReturn from types/agora.ts -> hooks/video/agora/useAgoraScreenShare.ts
